chore(Step): rename to ts
