+++ components/step/__tests__/a11y-spec.tsx
