+++ components/step/mobile/index.tsx
