+++ components/step/style.ts
